simple main.py
