+++ server/requirements.txt
@@ -7,4 +7,3 @@ sqlalchemy == 2.0.19
 alembic == 1.11.1
 passlib == 1.7.4
 pydantic == 2.1.1
-bcrypt == 4.0.1
